cz7host_frp/discord_bot/requirements.txt
discord.py
python-dotenv
httpx
fastapi
uvicorn[standard]